fix: remove old calendar
