--- src/Components/Calendar2/Calendar.module.scss
@@ -1,56 +1,0 @@
-.container {
-  --fc-button-bg-color: var(--cds-interactive-01);
-  --fc-button-border-color: var(--cds-interactive-01);
-  --fc-button-hover-bg-color: var(--cds-hover-primary);
-  --fc-button-hover-border-color: var(--cds-hover-primary);
-  --fc-button-active-bg-color: var(--cds-active-primary);
-  --fc-today-bg-color: var(--cds-ui-03);
-  position: relative;
-}
-
-:global .fc {
-  /* the calendar root */
-  margin: 0 auto;
-  tr {
-    background: white;
-  }
-}
-
-:global .fc-button-primary:focus {
-  box-shadow: inset 0 0 0 1px #0072c3, inset 0 0 0 2px #f2f4f8 !important;
-}
-
-.eventContentContainer {
-  align-items: center;
-  display: flex;
-  gap: 0.25rem;
-  color: var(--cds-text-01);
-  font-family: var(--cds-font-family);
-  font-size: 0.75rem;
-  max-width: 100%;
-  overflow: hidden;
-  svg {
-    min-height: 1rem;
-    min-width: 1rem;
-  }
-
-  &[data-status="trigger_disabled"] {
-    text-decoration: line-through;
-  }
-}
-
-.statusCircle {
-  &[data-status="active"] {
-    fill: var(--cds-support-success);
-  }
-  &[data-status="inactive"] {
-    fill: var(--cds-text-03);
-  }
-  &[data-status="deleted"] {
-    fill: var(--cds-support-error);
-  }
-
-  &[data-status="trigger_disabled"] {
-    fill: var(--cds-text-03);
-  }
-}
