--- src/Components/Calendar2/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Calendar";
