+++ src/Features/Schedule/Schedule.module.scss
@@ -34,9 +34,11 @@
 }
 
 .content {
-  height: calc(100vh - 10.3125rem);
+  height: calc(100vh - 10rem);
   padding-top: 1rem;
   width: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 .filtersContainer {
